requirements.txt
python-telegram-bot
aiohttp
iso18245